Move from stream namespace to io namespace
