+++ src/io/register_libvg_io.hpp
@@ -1,5 +1,5 @@
-#ifndef VG_STREAM_REGISTER_LIBVG_IO_HPP_INCLUDED
-#define VG_STREAM_REGISTER_LIBVG_IO_HPP_HPP_INCLUDED
+#ifndef VG_IO_REGISTER_LIBVG_IO_HPP_INCLUDED
+#define VG_IO_REGISTER_LIBVG_IO_HPP_HPP_INCLUDED
 
 /**
  * \file register_libvg_io.hpp
@@ -8,7 +8,7 @@
 
 namespace vg {
 
-namespace stream {
+namespace io {
 
 using namespace std;
