+++ src/io/register_loader_saver_gbwt.cpp
@@ -11,7 +11,7 @@
 
 namespace vg {
 
-namespace stream {
+namespace io {
 
 using namespace std;
 using namespace vg::io;
