+++ src/io/register_loader_saver_minimizer.hpp
@@ -1,5 +1,5 @@
-#ifndef VG_STREAM_REGISTER_LOADER_SAVER_MINIMIZER_HPP_INCLUDED
-#define VG_STREAM_REGISTER_LOADER_SAVER_MINIMIZER_HPP_INCLUDED
+#ifndef VG_IO_REGISTER_LOADER_SAVER_MINIMIZER_HPP_INCLUDED
+#define VG_IO_REGISTER_LOADER_SAVER_MINIMIZER_HPP_INCLUDED
 
 /**
  * \file register_loader_saver_minimizer.hpp
@@ -8,7 +8,7 @@
 
 namespace vg {
 
-namespace stream {
+namespace io {
 
 using namespace std;
